nfc-project/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pycryptodome
